requirements.txt
streamlit
streamlit-chat
pandas
Pillow
plotly
requests
transformers
python-dotenv